+++ src/rendering/cards/index.js
@@ -20,3 +20,4 @@ export { renderRoomCard } from './roomRenderer';
 export { renderCameraCard } from './cameraRenderer';
 export { renderSpacerCard } from './spacerRenderer';
 export { renderWaermepumpeCard } from './waermepumpeRenderer';
+export { renderLuftungsanlageCard } from './luftungsanlageRenderer';